Fix support for some schemas on query parameters
On the query parameters, the schema definition is unfolded into the set
of fields that compose it. This change ensures that only the relevant fields
on the schema records are passed in to the code that generates the
Swagger spec.

Before, the use of schemas that include more information than the model
definition itself on the query parameters was causing an
`IllegalArgumentException` to be raised.

Closes #104