Support java.time coercion
This adds support for java.time.Instant, java.time.LocalDate, and
java.time.LocalTime.

java.time extensions are inlined in the relevant namespaces, wrapped
with a macro that conditionally evaluates the body.

Refactored time coercion in coerce.clj to use multimethod dispatching.
this made it easier to extend conditionally for java.time.